pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fs-indexer"
version = "0.1.0"
authors = [
  { name="David Phillips" },
]
description = "High-performance file system indexer with SQLite backend"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "dbos",
    "sqlalchemy",
    "alembic",
    "xxhash",
    "pyyaml",
    "psycopg2-binary",
    "redis==5.0.1",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.3.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["fs_indexer"]
